Weird oxygen glitch with Zollus.
For some reason, when spawning certain mobs on Zollus, the game crashes
due to a NullPointerException when it tries to find a breathable block,
but I've directly told it that there is no breathable atmosphere on this
planet. Seems to only happen on this planet. There are a total of three
so far, only one of which is habitable. But Kriffus, a hot planet, does
not seem to have this glitch. I am completely stumped.