start to translate new items
It's not done yet, I might need some help in this.
First, the Shadow series seems would be a type of new element, But I can't understand what Shadow Spawner really is, It doesn't look like a egg of mob. More like a gate in the png image on github.
Then, witherite seems like a element just like other new elements. Is this more like ludacrite or a Essence like shadow essence or eden's essence?
It seems the new planet is corrupting by the shadow, maybe I should pay attention on it in furthur translation.
I will added the unfinshed stuff in next upload.